scraper-python/fbreaper_microservice/requirements.txt
kafka-python
playwright
fastapi
uvicorn
pydantic
requests
textblob
